Wrap text properly in the grid view of All files activity
Earlier the image icon size was quite large due to which a fraction
of text height was invisible even though the layout_height for the
text view was set to wrap_content. Now, the image icon dimensions have
been changed from 72dp to 64dp so that the text is wrapped properly
within the text view.