datamodel: Don't try to be smarter than the JVM by calling the garbage collector.
Calling System.gc() or Runtime.getRuntime().gc() is a bad idea for a simple reason: there is no way to know exactly what will be done under the hood by the JVM because the behavior will depend on its vendor, version and options:

Will the whole application be frozen during the call?
Is the -XX:DisableExplicitGC option activated?
Will the JVM simply ignore the call?

An application relying on these unpredictable methods is also unpredictable and therefore broken. The task of running the garbage collector should be left exclusively to the JVM.