authentication: do not set unlock timestamp when prompting for passcode.
Currently, every time an activity is stopped the `PassCodeManager` registers the current time and stores it as the last unlock time. This should not happen when the pass code activity is being stopped. To make matters worse, when the passcode activity is started whatever other activity was initially being started is then stopped which also registeres the last unlock time. By killing the app and restarting it before the 5 second timeout expires the lock will be bypassed entirely.

The only reason the lock works at all when not killing the app is simply because the PassCode activity is on top of the activity stack so the user is presented with it when resuming the app.

To combat the issue this commit introduces two changes.

1. Do not register unlock timestamp when the PassCodeActivity is stopped.
2. When the PassCodeActivity is started, set a `LOCK_PROMPTED` to `true`. When the PassCodeActivity is stopped we set the `LOCK_PROMPTED` to `false`.

Only when `LOCK_PROMPTED` is false and an activity other than the `PassCodeActivity` is stopped will we register the unlock time.

This fixes #3757

Signed-off-by: ardevd <edvard.holst@gmail.com>