Improve performance of DocumentsProvider#isChildDocument()
The old solution iterates through all parents of the child until it
finds the given parent or the storage root.
This has been show to be very expensive in empirical tests.

Therefore, this commit introduces a more efficient solution that simply
compares the file paths of child and given parent.
It also ensures that parent and child belong to the same account.

Reviewers need to take extra care that this change does not introduce
security issues by claiming a document is a child of a parent when it is
really not.

Signed-off-by: Torsten Grote <t@grobox.de>