Reset last sync date of files written via DocumentsProvider
The sync operation detects changes by comparing timestamps in
milliseconds. However, the local modification time only has a precision
of seconds (even though given in milliseconds).
When the DocumentsProvider is used programmatically, operations such as
a create and a subsequent write can happen within the same second
causing the sync operation to not detect the change and therefore
failing to upload a file causing data loss.

This commit also moves the file close listener to the thread used by the
DocumentsStorageProvider as this is more appropriate than the UI thread.

Fixes #6726

Signed-off-by: Torsten Grote <t@grobox.de>