Replace server and server.com with an invalid domain everywhere.
This should fix the timeout crashes in screenshot tests.
This also speeds up integration tests significantly, as the resolution fails
immediately, in contrast to the timeout for server.com.

As a bonus, we don't ping server.com from our IPs when running tests :)

I've chosen nextcloud.localhost for a fake domain, but suggestions are welcome.

Signed-off-by: Álvaro Brey <alvaro.brv@gmail.com>