CI:skip QA build if keystore pass is not available
This is just another step towards green checks on PRs :)

There is no way to skip the whole workflow depending on whether it is
a PR from a fork or not. The only workaround is to check whether secrets are present or not.

So, I've added a step to check if KS_PASS is empty, and set an output depending on that.
If it is empty, we just skip the rest of the steps.

To avoid the `if` repetition, a separate job can be used and then the whole `qa` job can
depend on its outputs. But that would add another "check" in the github UI.

Inspired by https://github.com/nextcloud/android/pull/8791#issuecomment-890524333

Signed-off-by: Álvaro Brey <alvaro.brv@gmail.com>