Fix behavior for KEEP_SERVER conflict resolution
Previously, choosing to keep the server version of a file in case of a
sync conflict triggered a download that overwrote the local version of
the file with the server version. This is unexpected as the sync
conflict dialog is only concerned with which file to keep on the
server.

Signed-off-by: BurningKarl <10158964+BurningKarl@users.noreply.github.com>