stale action: Add exempt label to mark issues that shouldn't be stale automatically
We've got some issues that are open since a long time ago but aren't resolved, and  there isn't some clear short term resolution to them.
As such these isues are kept open because they still affect users, but the users have to be "fighting" the stale bot, which
understandably generates confusion and frustration.

This PR adds a specific label we can add to those issues so that the stale action will ignore them.

Signed-off-by: Álvaro Brey Vilas <alvaro.brey@nextcloud.com>