When stopping PlayerService, do not use foreground service api
Sending async message to the service using foreground API
requires setting foreground flag. This won't happen if
the service is being stopped, so we must use startService()
API instead.

This API will throw IllegalStateException when app is banned
from running a service due to power management constraints.

Since stopping a service under such condition is not needed
anyway (it is not running), we can safely catch and ignore it.

This happens when app is woken up after long inactivity and
UI is recreated from cache.

Fixes #9491

Signed-off-by: Chris Narkiewicz <hello@ezaquarii.com>