Migrations: add migration to remove (invalid) media folder created by incorrect scan
This folder was created in version 3.18.1  - 3.20.2 (inclusive) when moving media files to private storage.

This is caused by MediaScanner not being able to access the private storage, and instead of crashing it creates
the (invalid) folder, with folders named the same as the original pictures.

Note: this will only work if the app has been granted storage permission, and the fake images won't disappear from
gallery apps / mediaStore until next reboot or media scan.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>