Prevent crash when selecting too many files to copy or move
This is caused by too much data in intent extras as we're passing a fat arraylist of objects.
As we only need the paths for the copy operation, change it so that only the paths are passed.

Ideally this would instead be stored in FDA or whatever listener we have for the folder picker,
rather than passed back and forth between the folder picker.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>