FilesUploadWorker: fix infinite loop when uploads failed
With the previous approach (uploadsStorageManager.gtCurrentAndPendingUploadsForAccount),
an upload that is, for example, waiting for wifi, would always be returned in the call,
thus the list would never be empty.

To avoid this but still process the uploads in pages to avoid OOM, we'll request them page-wise
from the StorageManager, with the pages ordered by ascending ID. This way we will only process each upload once,
but newly added uploads after we start will not be missed.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>
Signed-off-by: batpio <prbator@gmail.com>