Add testing utilities to override injections in classes injected with AndroidInjection
This includes Activities, Fragments, Services, etc. Other classes should use constructor injection,
and as such shouldn't need this.

This works by overriding the androidInjector method in the Application class (new TestMainApp added)
and allowing to use custom injectors there.

A InjectionOverrideRule is included for ease of use, but for more complicated cases, injection can be
controlled by accessing TestMainApp directly.

A sample is included in InjectionTestActivityTest.

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>
Signed-off-by: batpio <prbator@gmail.com>