Codecov: don't post status until 2 builds received
We currently upload unit test coverage and UI test coverage separately.
This tells Codecov to not compute/notify coverage until both are received.

This _should_ prevent the flopping coverage when one of the test suites fails,
or when CI is skipped (due to no code changes, for example, for Drone)

Signed-off-by: Álvaro Brey <alvaro.brey@nextcloud.com>