Extra options for uploading subfolders
Adds some extra options for uploading photos into various patterns of
subfolders. yyyy or yyyy/MM or yyyy/MM/dd supported.

Currently just adds an extra field for this, and has to be used in
conjunction with the "use subfolder" checkbox, the new option is greyed
out and disabled when "use subfolder" is not checked.

Signed-off-by: Dean Birch <dean.birch0@gmail.com>