enh/connectionChecks: Report main reason + don't do unnecessary checks
Previously checkConditions() would always report the last detected condition (in the method) that prevented the upload rather than the first. Besides doing unnecessary work, this meant that situations where say the auto-upload condition is set to "charging only" but network connectivity isn't working, would be reported as "Upload failed" due to a "Connection error" which is irrelevant. Also means we stop doing silling things like checking for network connectivity when the upload is already excluded due to the device not currently charging.

Signed-off-by: Josh Richards <josh.t.richards@gmail.com>