Move or Copy files: Start in the parent folder instead of root
Previous behavior always moved up to the root folder, which meant the user
lost the context of their current files. Especially in complicated nested
folder structures, this made standard operations overly complicated, like
duplicating in the same folder or duplicating/moving to a sibling folder.

This is now also in sync with the behavior of the iOS app and the web UI

Signed-off-by: Philipp Hasper <vcs@hasper.info>