Fixing annotation for optional OCFile parameter
Changing to OCFile? is required, as the parameter is indeed nullable.
This didn't turn up prior to the rebase nor was it flagged in the
pre-existing code, even though it is already an issue there.
What probably happened is a combination of updated detection tools and
the tools purposefully only flagging changed lines.

Additionally, the nullness annotation was added to the underlying Java
calls, to make sure the Kotlin type definition holds true.

Signed-off-by: Philipp Hasper <vcs@hasper.info>