Asserting intent key non-nullness at time of definition
The canonicalName could only be null, if the class is hidden, a local class
or an an anonymous class. This is not the case at hand.
Also, if any of these keys is null, all of the keys would be null, as they
are constructed in a similar pattern. And then they would be
indistinguishable from each other, negating the purpose of keys.

Signed-off-by: Philipp Hasper <vcs@hasper.info>