Fixing timing issue with FileActivity's loading dialog show and dismiss.
Before dismissing the dialog, we need to wait for a potentially pending
transaction. As showing the dialog also includes the dismissing of prior
instances, we need to wait there as well.
Both is needed to satisfy the test case added in the previous commit.

Otherwise, the dialog might be shown after it was meant to be dismissed
already. This issue was observed when testing RemoveFilesDialogFragment's
removeFiles() and also sporadically "in the wild".

Signed-off-by: Philipp Hasper <vcs@hasper.info>