When uploading a single file, offer input field to change file name
This streamlines the upload from files, e.g. from email attachments.
Closes #9215.
The approach via the listener mFileDisplayNameTransformer was chosen
because this allows a potential extension to multiple files (with multiple
input fields). Alternatively, ReceiveExternalFilesActivity#uploadFile()
could've been used, which requires the storage to a temporary file
first.

nextFocusForward and nextFocusDown are both needed to support navigation
with the Tab key as well as the Enter key

Signed-off-by: Philipp Hasper <vcs@hasper.info>