Verify file name validity during user input.
The automated test only tests this partially, because the only invalid
name was an emtpy name. All other characters which are typically unsuitable
for a filename, like the directory separator '/' are not properly rejected
during the test run, because the FileNameValidator is based on the user
capabilities, which are null because this is a sever-less test.

Signed-off-by: Philipp Hasper <vcs@hasper.info>